Move publish button and functions inside QuestionnaireCreate
